+++ NEWS.md
@@ -1,5 +1,7 @@
 # ipfr (development version)
 
+  * Fixed bug when marginal target only had 1 category (#2)
+
 # ipfr 1.0.0 (2019-08-21)
 
   * First release of the package.